doctor-appointment-backend/requirements.txt
flask 
flask-sqlalchemy 
flask_migrate 
flask-jwt-extended 
python-dotenv
flask_cors
supabase
uvicorn
fastapi

